Rename magic portfolio components to dynamic capital
