+++ apps/web/app/about/page.tsx
@@ -13,8 +13,8 @@ import {
   Row,
 } from "@once-ui-system/core";
 import { baseURL, about, person, social, toAbsoluteUrl } from "@/resources";
-import TableOfContents from "@/components/magic-portfolio/about/TableOfContents";
-import styles from "@/components/magic-portfolio/about/about.module.scss";
+import TableOfContents from "@/components/dynamic-capital/about/TableOfContents";
+import styles from "@/components/dynamic-capital/about/about.module.scss";
 import React from "react";
 
 export function generateMetadata() {
