+++ apps/web/app/layout.tsx
@@ -23,7 +23,7 @@ import {
   Header,
   RouteGuard,
   ScrollToHash,
-} from "@/components/magic-portfolio";
+} from "@/components/dynamic-capital";
 import { dynamicBranding, dynamicUI } from "@/resources";
 
 const SITE_URL = process.env.SITE_URL || "http://localhost:8080";
