+++ apps/web/app/plans/page.tsx
@@ -1,7 +1,7 @@
 import { Heading, Column, Text } from "@once-ui-system/core";
 
-import { VipPackagesSection } from "@/components/magic-portfolio/home/VipPackagesSection";
-import { CheckoutCallout } from "@/components/magic-portfolio/home/CheckoutCallout";
+import { VipPackagesSection } from "@/components/dynamic-capital/home/VipPackagesSection";
+import { CheckoutCallout } from "@/components/dynamic-capital/home/CheckoutCallout";
 
 export const metadata = {
   title: "VIP Plans – Dynamic Capital",
