+++ apps/web/components/dynamic-capital/Footer.module.scss
