+++ apps/web/components/dynamic-capital/Footer.tsx
