+++ apps/web/components/dynamic-capital/Header.module.scss
