+++ apps/web/components/dynamic-capital/Header.tsx
