+++ apps/web/components/dynamic-capital/HeadingLink.module.scss
