+++ apps/web/components/dynamic-capital/HeadingLink.tsx
