+++ apps/web/components/dynamic-capital/Mailchimp.tsx
