+++ apps/web/components/dynamic-capital/ProjectCard.tsx
