+++ apps/web/components/dynamic-capital/about/about.module.scss
