+++ apps/web/components/dynamic-capital/blog/Post.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Card, Column, Media, Row, Avatar, Text } from "@once-ui-system/core";
-import { formatDate } from "@/utils/magic-portfolio/formatDate";
+import { formatDate } from "@/utils/dynamic-capital/formatDate";
 import { person } from "@/resources";
 
 interface PostProps {
