+++ apps/web/components/dynamic-capital/breakpoints.scss
