+++ apps/web/components/dynamic-capital/gallery/GalleryView.tsx
