+++ apps/web/components/dynamic-capital/home/AboutShowcase.tsx
