+++ apps/web/components/dynamic-capital/home/ComplianceCertificates.tsx
