+++ apps/web/components/dynamic-capital/home/FundamentalAnalysisSection.tsx
