+++ apps/web/components/dynamic-capital/home/FxMarketSnapshotSection.tsx
