+++ apps/web/components/dynamic-capital/home/HeroExperience.tsx
