+++ apps/web/components/dynamic-capital/home/MarketWatchlist.tsx
