+++ apps/web/components/dynamic-capital/home/PerformanceInsightsSection.tsx
