+++ apps/web/components/dynamic-capital/home/PoolTradingSection.tsx
