+++ apps/web/components/dynamic-capital/home/ValuePropositionSection.tsx
