+++ apps/web/components/dynamic-capital/home/VipPackagesSection.tsx
