+++ apps/web/components/dynamic-capital/index.ts
