+++ apps/web/components/dynamic-capital/mdx.tsx
