+++ apps/web/components/dynamic-capital/work/Projects.module.scss
