+++ apps/web/utils/dynamic-capital/formatDate.ts
