+++ apps/web/utils/dynamic-capital/utils.ts
